requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
python-dotenv